Reorganized urql client
